feat: add guardian settings
