+++ apps/web/src/pages/settings/danger.tsx
@@ -1,0 +1,3 @@
+import DangerSettings from '@components/Settings/Danger';
+
+export default DangerSettings;
