--- apps/web/src/pages/settings/delete.tsx
@@ -1,3 +1,0 @@
-import DeleteSettings from '@components/Settings/Delete';
-
-export default DeleteSettings;
